+++ core/database.php
@@ -2,14 +2,14 @@
 # Database
 namespace Core;
 
-use PDO;
+use \PDO;
 
 class Database extends PDO
 {
 	public function __construct($DB_TYPE, $DB_HOST, $DB_NAME, $DB_USER, $DB_PASS)
 	{
         try {
-            parent::__construct($DB_TYPE . ':host=' . $DB_HOST . ';dbname=' . $DB_NAME, $DB_USER, $DB_PASS);
+            parent::__construct($DB_TYPE . ':host=' . $DB_HOST . ';dbname=' . $DB_NAME . ';charset=utf8', $DB_USER, $DB_PASS);
             parent::setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
         } catch(PDOException $ex) {
             echo '<h3><br />Catch Connect Error--->>> ' . $ex->getMessage() . '<br /></h3>';